fix path (#30)
