--- src/pages/about.tsx
@@ -1,7 +1,0 @@
-import type { NextPage } from "next";
-
-const About: NextPage = () => {
-  return <p>Hello Items</p>;
-};
-
-export default About;
